requirements.txt
fastapi
uvicorn
langchain
chromadb
tiktoken
python-dotenv
numpy
langchain-openai
langchain-community
langchain-chroma
openai